top(1): Handle SCORE (dumping) process state similar to other states
Before this change, the SCORE process state would shown as '?P/5' in the
'STATE' column, which could be confusing.

Similar to existing process states (e.g., running, zombie), handle the
SCORE (dumping a core) process state as well.  This state will show in
the process line in the header, and also shown as 'DUMP' in the 'STATE'
column for such processes.

Update the man page accordingly.  In addition, add copyright text to the
man page.

Reported-by: kworr