installer(8): Distinguish setting and getting of encryption passphrase
The installer was using the *same* dialog to set and get the passphrase
for filesystem encryption/decryption, which was confusing and a bit
inconvenience (because the passphrase was required to input twice for
decryption).

The decryption passphrase was get in the following two places:
1. Just after installation and begins to configure the system, the
   passphrase is asked to decrypt and mount the filesystems;
2. Use the LiveCD to "Configure an installed System".

This commit improves the fn_get_passphrase() function to distinguish the
setting and getting of the encryption passphrase.  The dialog of setting
the passphrase remains the same, but the dialog of getting the
passphrase is simplified and doesn't require to confirm the input.

This commit is derived from the patch by tuxillo.

Bug: https://bugs.dragonflybsd.org/issues/3028