libc/ukp_setproctitle: Properly handle skipping the executable's name.
If fmt begins with a "-" character, the executable's name should be
skipped. Our fast version of setproctitle() - which is used after the
10th call to setproctitle() - was not handling that correctly.

While here, fix the case where NULL is passed to setproctitle(), which
restores the process title to its original value. Just defer back to
setproctitle() in this case. Thanks to aly for catching that.

Dragonfly-bug: <https://bugs.dragonflybsd.org/issues/3319>