boot: Expose SMBIOS entry point via kenv
The SMBIOS Reference Specification (v3.6.0) says (in Section 5.2.1):

"On UEFI-based systems, the SMBIOS Entry Point structure can be located
by looking in the EFI Configuration Table for the SMBIOS GUID
(SMBIOS_TABLE_GUID, {EB9D2D31-2D88-11D3-9A16- 0090273FC14D}) and using
the associated pointer."

So export the SMBIOS entry point via kenv to allow the SMBIOS table be
found in UEFI-only mode, where the BIOS is not mapped into the standard
address space and thus the SMBIOS table might not appear between 0xf0000
and 0xfffff.  For example, dmidecode(8) was failing with:

    Scanning /dev/mem for entry point.
    # No SMBIOS nor DMI entry point found, sorry.

Use the same kenv 'hint.smbios.0.mem' as FreeBSD to help with dports(7).

Refer-to: FreeBSD (revision 307326)
Discussed-with: swildner