net/radix: Fix rn_refines() and add the function description
After 'n++' called in the 'longer' assignment, the clen(n) becomes
invalid since the first byte is no longer the key length.  This caused
the function to always return FALSE and could lead to out-of-bound
memory access.

Fix the bug, and describe the function's usage and return value.

The bug was introduced in commit 2e9572d in 2004.

Reported-by: dczheng