sbin/hammer2: Fix -Werror=format-truncation= warning
The first snprintf is formatting a potentially longer string into
a smaller buffer. Expanding filename[] size to larger than pfs.name[]
fixes a format-truncation warning, but that generates a second
format-truncation warning on formatting the string back to pfs.name[]
in the second snprintf. The second snprintf can be done with strlcpy.