sys/vfs/hammer2: Remove unused lbase local variables for hammer2_calc_logical()
These variables are never used after assigned values.

If the third argument is NULL, the second argument is unused either,
in which case 0 is usually passed. If the fourth argument is NULL,
the first argument is unused either, in which case NULL can be passed.

(So if at least the last 2 arguments are NULL, caller is just looking
for an fs block size, which is always 64KB...)