csprng: Update to use crypto/chacha20 (a better version)
The CSPRNG code was already using Chacha20 (from crypto/chacha) to
generate the random stream.  However, the 'crypto/chacha20' version
has been tweaked for and better suited to random stream generation.
The enhancements include:

- Provide the KEYSTREAM_ONLY mode to help ease the invocation and
  improve performance.
- Allow to use a 128-bit counter to avoid worrying about overflow
  (i.e., nonce reuse).  This also remove the burden on the caller
  to check for counter overflow and rotate nonce.
- Can be embedded for better compiler optimization.

The 'crypto/chacha20' was imported on 2023-02-25 from FreeBSD to
implement the libc arc4random(3) API.  After the CSPRNG migration,
the old 'crypto/chacha' version become unused and will be removed
in a later commit.

Referred to FreeBSD and OpenBSD.