Fix 4 uses of PF_ANEQ
The uses were depending on the old broken behavior of always checking
against IPv6 unless the family was specifically AF_INET. In 2 cases, the
AF argument was 0 seeking for a full comparison. Pass AF_INET6 instead
for these two cases to restore the intended behavior of a full
comparison. The other two cases appear to just be uncaught typos. They
are inside of an AF_INET6 case, and there is no obvious reason why you
would want an AF_INET comparison in this case.