net/radix: Document rn_inithead() and change to use byte offset
The radix code only supports offset bit being multiple of bytes but not
any byte fraction, so it's clearer and better to just use byte offset
instead of bit offset.  In addition, the caller can be easier to
determine this argument using offsetof().

On the other hand, ignore this off_bytes parameter and auto set it to
zero when to create a mask tree.

Also add a detailed description for rn_inithead().