net/radix: Implement rn_flush() and rn_freehead()
rn_flush() flushes all nodes in the given radix tree and thus makes the
tree empty; rn_freehead() can then free the emptied tree.

The separation of flushing and freeing is required to support the use
case in ipfw2, which needs to flush a table without destroying it.

A mask tree is somewhat special, because its nodes are allocated inside
this radix code, so the rn_freemask() callback function is provided to
be used in rn_flush() to flush a mask tree and reclaim its memory.

These APIs can help the radix users to easily flush and free the radix
trees without rolling their own versions.