wg: Remove the unused 'sc_ucred' field and related code
This was used in FreeBSD to control the permissions of an WG interface
in an VNET, in order to achieve something similar to Linux's network
namespace.

However, DragonFly doesn't yet support this feature, so the 'sc_ucred'
field becomes unused and thus can be removed.  It's easy to get it back
when we're ready to implement this feature in the future.