ifconfig.8: Remove the legacy and commented eon(5) and OSI stuffs
The eon(5) seems to be related to RFC 986, "which appears to be a rough
ISO equivalent of the IPv6 6to4 transition mechanism." [0]

The "nsellength" thing seems to be related to OSI networking.

These commented contents has been there since 1996 in FreeBSD codebase.
I don't see a point to still keep them, so just remove these legacy
stuffs.

[0] https://gnats.netbsd.org/10500