mbuf(9): Improve m_copyback2() to use the trailing space
Make use of the trailing space of the last mbuf if available, both in
skipping the offset and in copying back the data.

As a side benefit of this change, the 'm0->m_len' field no longer needs
to be initialized and can simply be 0, as in a newly allocated mbuf.

Referred to both OpenBSD and FreeBSD; however, they only try to use the
trailing space in copying the data.