crypto/chachapoly: Allow output be NULL in decrypting empty plaintext
Don't distinguish the input cipher data from AD by checking whether the
output buffer is NULL, because it's actually valid to pass it as NULL
when to decrypt a message of empty plaintext.  And it's really used by
WireGuard.

So separate the AD process code into a separate helper function named
_chacha20poly1305_update_ad().   Update the assertions to not blindly
assert 'out != NULL'.  Also add a note about this special case to the
header file.