wg: Clean up noise_keypair_counter_check() a bit
Remove '++recv' together with the '+ 1' calculation, so the code nows
becomes more understandable, and the conditional of
'kp->kp_counter_recv >= REJECT_AFTER_MESSAGES' also becomes the same as
the one in noise_keypair_decrypt(); so it reduces the confusion between
them.  The selftest is also passed.  (Referred to OpenBSD)

In addition, add brief comments to describe the 'kp_counter_send' and
'kp_counter_recv'.