wg: Add RXCSUM support to avoid unnecessary checksum validation
The packet that is about to be delivered in is authentic as ensured by
the AEAD tag, so we can tell the networking stack that this packet has
valid checksums and thus is unnecessary to check again.

Therefore, implement RXCSUM support for the wg interface, and update the
ioctl() to support to enable/disable this feature.

Meanwhile, move the mbuf flags clearance code just before the delivery,
i.e., netisr_queue() and wg_send().