wg: Write rc(8) script to easily manage wg(4) interfaces
This "wg" rc(8) script is somewhat similar to the "wg-quick" tool on
Linux/FreeBSD.  It can be used to quickly start/stop the wg(4)
interfaces according to the wg.conf(5) configuration files in the
"/etc/wireguard" directory.

The syntax of wg.conf(5) configuration file is very similar to that
of "wg-quick" but with necessary changes and minor additions.  See
wg.conf(5) for details.

On the one hand, the new "wg_enable" and "wg_interfaces" variables in
"/etc/rc.conf" can be used to auto-configure the wg(4) interfaces during
the system startup.  See rc.conf(5) for more details.

On the other hand, this "wg" script can be manually called from the
command-line to start/stop the wg(4) interfaces.

Thanks to swildner for reviewing the man page.