rc.d: Add FILESYSTEMS dummy dependency
This dummy dependency ensures that root and other critical local file
systems are mounted.  One of the intentions to add this dummy dependency
is to fix the warnings caused by some ports' rc script wrongly dependent
on this script.

Update the rc.8 man page; meanwhile, tweak the description of other
dummy dependencies a bit.

Updating existing rc scripts to make use of FILESYSTEMS is coming in
another commit.

Obtained-from: FreeBSD