ping(8): Allow normal users to specify larger packets
Before this change, only super-user could specify a packet size larger
than the default 56 bytes.  This restriction was added by Matt Dillon in
1998 during the BEST days [0].

Nowadays, it doesn't make much sense to limit the packet size.  In
addition, this restriction doesn't exist in ping6(8) and on NetBSD,
OpenBSD and Linux.

And one more interesting thing, UMS [1] uses this feature to estimate
the client's bandwidth to optimize the streaming experience.

[0] https://github.com/freebsd/freebsd-src/commit/526f06b278d9252add168aa18b60242c08771165
[1] UMS: https://github.com/UniversalMediaServer/UniversalMediaServer

Discussed-with: dillon