ping(8): Relax the interval restriction for normal users
The ping interval was restricted to be >= 1 second for normal users to
prevent from performing DoS attacks.  This restriction was added by
Matt Dillon in 1998 during the BEST days [0].  It has become less
meaningful now.  Following the Linux behavior, relax the interval to be
>= 2 milliseconds for normal users.

In addition, add a check for negative interval numbers.

[0] https://github.com/freebsd/freebsd-src/commit/526f06b278d9252add168aa18b60242c08771165

Discussed-with: dillon