sockstat(1): Fix and improve output formatting
The main issue with the old formatting was that the FD column was
limited to 2 characters, so there would be no space between FD and PROTO
columns.

Increase the FD column to 5 characters, and increase the PID column to
6 characters.  In addition, refactor the output formatting code to clean
up the mess.

While there, clean up the code a bit to save some type castings.