Fix null dereference bug in fuse and re-enable it.
There was a null dereference bug in fuse if the subtype was null.
subtype is supposed to be optional, as demonstrated by the code, so test
it for null before copying it in from userspace.

Remove the redundant memcpy which appears to have been missed when the
code was changed to memset/copyinstr.