net: Integrate link_rtrequest() into rt_getifa()
When a route was inserted but had an invalid IFA (rt_ifa), the
link_rtrequest() would be called to fix it with a link-level gateway.

Move the link_rtrequest() logic into the pre-insertion rt_getifa() step,
and thus simplifies the IFA selection.  This will also help the
forthcoming mangement of loopback routes for interface addresses.

Obtained-from: FreeBSD
- https://reviews.freebsd.org/D20076
- https://github.com/freebsd/freebsd-src/commit/2ad7ed6e4a84448234e9aaef3c9fa884848f7387