netinet: Improve the deletion of loopback route
When the target host address is still bound to another interface address
(e.g., an address on another interface added later), the loopback route
should be updated to point to that address instead of simply being
removed.

Given that our rtrequest1() only supports route addition and deletion,
but not route change, we have to first remove the original loopback
route and then re-add a route with the wanted interface address.

See also in_localip_more() and ifa_switch_loopback_route() in FreeBSD.