acpi: Narrow workaround for broken interrupt settings
Commit ad52dd3206 ("Sync acpi_resource with FreeBSD") introduced a
workaround for some broken BIOSes that specified active-lo instead of
active-hi polarity for ISA IRQs for UARTs.  The workaround assumed that
edge-sensitive ISA IRQs on x86 should always be active-hi.  However,
some recent AMD systems actually use active-lo edge-sensitive ISA IRQs
(and not just for UARTs, but also for the keyboard and PS/2 mouse
devices) and the override causes interrupts to be dropped resulting in
boot time hangs, non-working keyboards, etc.

Add a hw.acpi.override_isa_irq_polarity tunable (readable as a sysctl
post-boot) to control this quirk.  It can be set to 1 to force enable
the override and 0 to disable it.  The log of FreeBSD-PR (see below)
mentions an Intel motherboard as the sample case, so default the
tunable to 1 on systems with an Intel CPU and 0 otherwise.

Special thanks to Matthias Lanter <freebsd@lanter-it.ch> for tracking
down boot time issues on recent AMD systems to mismatched interrupt
polarity.

Taken-from: FreeBSD [1]

FreeBSD-PR: https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=270707

[1]: https://github.com/freebsd/freebsd-src/commit/0a34d050ae8ea14feddd3d2a62fd2f612613b2c5