initrd: Add diff/sdiff back to rescue and fix build
Update the rescue/Makefile to include the updated diff/sdiff utilities;
however, diff3 is excluded because it's a shell script and relies on
libexec/diff3prog.  As a result, we must remove /rescue/diff3 via
"make upgrade"; otherwise, /rescue/diff3 would become a separate binary
(because of hardlinks) and break initrd build because of exceeded size.

Bug: https://bugs.dragonflybsd.org/issues/3376