nrelease: Fix volume label to comply with cd9660 spec
The iso/cd9660 spec says the volume identifier/label can have a maximum
length of 32 characters, and must contain only A-Z, 0-9 and _, so we
have to fix the label portion of git version string to meet the
requirement.

Although the old label format is accepted by mkisofs, it will be
rejected by makefs(8).  So this change helps we migrate to the latter.