route(4): Ensure RTA_NETMASK has a valid length and family
When adding a route to the radix tree only the data part of the
sockaddr is preserved. Trailing zeros are also removed which
results in zero length netmask for the default route.
So when we pull the route from the kernel via a route(4) message,
the netmask sockaddr has an invalid family and for the default
route, an invalid length as well.

You can see the effect of this by ensuring a default route is
present and then issuing the command `route -n flush`
which results in an error.

The simple fix is this: when we place the netmask in a route(4)
message and we have a destination address (ie it's a route)
then we can fix the family (it will be either 0x00 or 0xff) and
the length (it will be 0 for the default route).