procctl(2): Implement PROC_REAP_KILL request to send signals
The PROC_REAP_KILL request can be used to deliver a signal to the direct
children or all descendants of the reaper.  For the latter case, the
signal will not be delivered to the descendants of any sub-reapers.

The similar request has been implemented in FreeBSD for many years, and
has been used in their timeout(1).  Actually, our timeout(1) is
misbehaving if the '--foreground' flag is not specified.  And that's one
of the main reasons to implement this request.

Mentored-by: dillon
Reviewed-by: dillon