timeout(1): Fix signal delivery bug by using PROC_REAP_KILL
When the --foreground | -f option is not specified, timeout(1) should
send the signal to all the descendants.  As we're using the procctl(2)
reaper framework, we were lacking the PROC_REAP_KILL request as in
FreeBSD that allows us to send the signal to all the descendants.

Now we've implemented the PROC_REAP_KILL request in procctl(2), so we
just use it and fix the signal delivery bug.