wg: Fix callout leak by adding callout_terminate()
DragonFly allocates an internal '_callout' struct for a 'callout' struct,
so the callout_terminate() function must be called to free the internal
struct before releasing a 'callout' struct.

On the contrary, FreeBSD fully embeds the info into a 'callout' struct
and does not need (and actually does not have) the callout_terminate().