procctl(2): Improve PROC_REAP_KILL to handle reparenting race
During the process tree iteration, a process or a process subtree might
be reparented (because of SIGTERM/SIGKILL or ptrace/debug).  In order to
handle that race, we reset the p_reaptid to 0 whenever a process gets
reparented, and then check the p_reaptid value during the iteration.  If
a reparent is detected, we restart the iteration.

Suggested-by: dillon
Reviewed-by: dillon