rc.d/root: Add 'nojail' keyword to fix jail boot
A jail cannot remount the root filesystem so it was failing to boot
because of error:

	Mounting root filesystem rw failed; startup aborted.

Add the 'nojail' keyword to exclude this rc script in jail boot.
FreeBSD also has this keyword for this script for 20+ years.

With this fix, a jail boots OK but there are still some errors during
the boot, so there are more rc scripts that need the 'nojail' keyword
or need tweaks for jail.  Will look into this later.

Reported-by: fgudin (Francis GUDIN) on IRC