crypto - Move core AESNI implementation to sys/crypto/aesni
AESNI is a set of CPU instructions that help to accelerate the AES
cipher. It is available on almost any x86 CPU since about 2010.

Moving this makes the code (and as such the AESNI instruction) usable
from outside the opencrypto framework while keeping aesni(4) from
dev/crypto/aesni still operational.