dm_target_crypt_ng - test compatibility of old and new module
This script tests compatibility between dm_target_crypt and
dm_target_crypt_ng.

It creates an encrypted disk with one implementation, then copies
files to it. Then it mounts it using the other implementation and
verifies that the copied files are still the same (using mtree).

It does this procedure in both directions and various configurations
(e.g. with or w/o AESNI enabled).