nrelease: Check GITREV and just fail if empty
The GITREV variable is determined by the tools/gitrev.sh script, which
might fail if the current user has insufficient permission to access the
source directory.  That would cause the `uname` to output something
strange that could confuse people.  So I think it's better to just fail
the build and require a fix to the build environment.

Reported-by: JustinS