diff(1): Update header time format to be compatible with GNU diff
The old code was calling ctime() to format the time, which had a format
like "Tue May  6 19:00:26 2025" that was different from the GNU diff.
Fix the code to format the time in the same format as GNU diff, e.g.,
"2025-05-06 19:00:26.113084000 +0800".

FreeBSD also made such a change to its diff.  However, we don't consider
the '-c' option here, because I checked with GNU diff and it didn't
change the time format with the '-c' option.

Referred-to: FreeBSD