fdisk(8): Calculate ncylinders and warn about mismatch
The virtio_blk(4) disk driver was directly reporting the ncylinders
parameter specified by the QEMU emulator, which however is capped at
16383 (the legacy MBR maximum).  That confused fdisk(8) to recognize the
disk to have only ~8GB size instead of the full capacity.

The issue virtio_blk(4) has been fixed in a former commit.  This commit
improve fdisk(8) by calculating the ncylinders on its own and comparing
it with the one provided by the disk.  If there is a mismatch, emit a
warning message to remind a fix to the disk driver.

Derived from the patch submitted by mikdusan in bug #3369.

Bug: https://bugs.dragonflybsd.org/issues/3369