atkbd: Fix first keystroke force reset
Some i8042 falsely return KBD_ACK for ECHO command which cause the
keyboard echo test to fail.
Without passing echo test, the keyboard is considered as unconfigured.

Although this is incorrect behavior, we still consider it correct to
prevent the forced reset (a step to configure a keyboard) of the entire
keyboard when the first key interrupt is reached.

This should fix keyboard probing at least for Framework Laptop [1].

Taken-from: FreeBSD [2] (minus typos)

[1]: https://reviews.freebsd.org/D50498
[2]: https://github.com/freebsd/freebsd-src/commit/cba5d7ab32fd0b4e9c4312cc967baf137377f812