syscons - add sysctl to disable console bell globally
Add sysctl `kern.syscons.enable_bell`, which globally disables the
system console bell ("sc_beep") when set to 0. By default the bell is
on.

To silence all beeps, add this to /etc/sysctl.conf:

    kern.syscons.enable_bell=0

This has the same effect as tcsh's "set nobeep". But there is no such
setting to my knowledge for our "/bin/sh".