nvmm: Don't require <machine/specialreg.h> and be more self-contained
Define the necessary CR/XCR/MSR registers and their bits by itself, so
that <machine/specialreg.h> is no longer used and any collisions are
avoided.  This is similar to the various CPUID defines and it makes NVMM
more self-contained and portable.

In addition, this commit includes the following improvements/cleanups:
- Separate the Intel- and AMD-specific registers and define them in
  their own implementations.
- Define 'MSR_CMPHALT' for AMD/SVM, avoiding the hardcoded 0xC0010055.
- Rename 'MSR_BIOS_SIGN' to 'MSR_IA32_BIOS_SIGN_ID' and
  'MSR_MISC_ENABLE' to 'MSR_IA32_MISC_ENABLE', in order to better align
  with the Intel SDM.