<sys/cdefs.h>: Add '__inline' to '__always_inline'
It's clearly the intention of the '__always_inline' macro applied to a
function to inline the function.  However, GCC seems to only allow the
'__attribute__((__always_inline__))' to be applied to inline functions;
in other words, function declarations with the attribute should also be
declared as inline.

Both GCC/Clang sources themselves use the two in combination:
	inline __attribute__((__always_inline__))

At this moment, DragonFly kernel souces (except the DRM sources ported
from Linux) also use the two in combination, i.e., the same as GCC/Clang
above.  So the main intention of this change is to help future code
ported from FreeBSD and other systems.

This change could helps eliminate the GCC warnings of this form:
warning: 'always_inline' function might not be inlinable [-Wattributes]

Obtain-from: FreeBSD
- Commit: d25f0bdceb3abb798342821791d7ececb4ba3411
- Differential Revision: https://reviews.freebsd.org/D45711

Discussed-with: swildner