gpt(8): Change '-p' to be create/init/migrate command option
The '-p' option was a generic and applied to all commands, but actually
only the "create", "init", "migrate" commands used it.  So just change
it to be a command-specific command for these three commands.

In order to avoid conflict, rename the original '-p' option to '-P' for
the "create" command.  This shouldn't cause any trouble, because this
option had been documented for debug purpose use.

Referred-to: NetBSD