bsd.{sys,kern}.mk: Fix syntax error in pattern 'gcc[5-]*'
If ${_WCCVER} doesn't match the 'gcc[5-]*' pattern, bmake would generate
this fatal error:

```
Unfinished character list in pattern 'gcc[5-]*' of modifier ':M'
	while evaluating variable "_WCCVER" with value "gcc100"
make: Fatal errors encountered -- cannot continue
```

Interestingly, this pattern works as expected if there is match.  This
is because bmake matches the pattern along the parsing; it doesn't
validate the pattern syntax before matching it.  The bmake code requires
that both the range begin and end be specified.  So this is actually a
bug in the bmake code.  See also contrib/bmake/str.c:Str_Match().

Reported-by: sparlane (IRC)
Discussed-with: swildner