test: Use snapshot testing for cryptoapi/cryptodev
The tests I wrote for cryptoapi where based on comparing the results
against the results of cryptodev - if they match, the test succeeds. But
as I am planning to remove cryptodev, this test methodology no longer
works.

Instead, store the test input data in fixtures.csv and run the test case
for each input. The generated output should then match the
expected_output.md in order for the tests to pass. This simplifies
adding new test fixtures and allows for re-generating the
expected_output.md.