test: Integrate benchmark into cryptoapi_test util
Print execution times and add "-n REPETITIONS" option.

Take the benchmark results with a grain of salt when comparing the
cryptoapi results against the cryptodev results, as we do not use
the AESNI CPU instruction when crytoapi is used outside the kernel.

While there, parse CLI options using getopt(3) and add a few more
test fixtures.