opencrypto - remove in-kernel crypto(9) framework
The opencrypto crypto(9) API was quite complex (8k LoC), slow and not
used by any other kernel subsystem within DragonFly anymore. It allowed
for chaining various operations, crypto ops and compression, but AFAIK
this was never really used and rather complicated.  For a much simpler,
synchronous API see sys/src/crypto/cryptoapi. For a nice writeup on the
problems of crypto(9) in the context of FreeBSD, please see [1].

The opencrypto API was asynchronous by design. This was good back in the
days when dedicated hardware crypto devices did exist to help offload
the CPU. But the world has now changed towards synchronous CPU
instructions like AESNI. These dedicated CPU instructions are best
called synchronously, which removes the need for book-keeping of
asynchronous requests, often greatly simplifying the caller as well as
the API and "backend" implementation.

Furthermore, those dedicated crypto CPU instructions are not limited to
the kernel, they can be directly used by userland applications as well.
OpenSSL does that for instance (in case the CPU instruction is
supported) and as such is independent from the OS and it's crypto
interface.  This further avoids context switches from userland to kernel
and vice versa. Also, userland applications can more easily be compiled
with -O3 compiler optimizations and this makes a huge different in
performance, especially when falling back to a "software" crypto
implementation.

For a much simpler and more limited, but synchronous crypto API that
both works in-kernel as well as in userland (without the need to go
through /dev/crypto device), see src/sys/crypto/cryptoapi. But usually,
you want to directly use a given block cipher without any abstraction on
top.

[1]: https://papers.freebsd.org/2022/bsdcan/baldwin-refining_freebsds_kernel_crypto_framework/